+++ Source/XmlParser/AplicacionParser/AplicacionParser/Grilla.h
@@ -5,16 +5,16 @@
 
 class Grilla
 {
-	size_t ancho;
-	size_t alto;
+	int ancho;
+	int alto;
 	std::string tipoObstaculoPorDefecto;
 	List<List<Celda>> matriz;
 	bool generarMatriz(List<XmlElement> le);
 public:
 	Grilla(XmlElement e);
-	Grilla(size_t an, size_t al, std::string topd);
-	size_t getAncho();
-	size_t getAlto();
+	Grilla(int an, int al, std::string topd);
+	int getAncho();
+	int getAlto();
 	std::string getTipoObstaculoPorDefecto();
 	List<List<Celda>> getMatriz();
 	~Grilla(void);
